+++ tests/test_core.py
@@ -27,8 +27,8 @@ from tools import shared, building, config, webassembly
 from common import RunnerCore, path_from_root, requires_native_clang, test_file
 from common import skip_if, needs_dylink, no_windows, is_slow_test, create_file, parameterized
 from common import env_modify, with_env_modify, disabled, node_pthreads
-from common import read_file, read_binary, require_node
-from common import NON_ZERO, WEBIDL_BINDER, EMBUILDER, EMMAKE
+from common import read_file, read_binary, require_node, require_v8
+from common import NON_ZERO, WEBIDL_BINDER, EMBUILDER
 import clang_native
 
 # decorators for limiting which modes a test can run in
@@ -77,6 +77,7 @@ def all_engines(f):
   def decorated(self):
     old = self.use_all_engines
     self.use_all_engines = True
+    self.set_setting('ENVIRONMENT', 'web,node,shell')
     try:
       f(self)
     finally:
@@ -463,7 +464,7 @@ class TestCoreBase(RunnerCore):
   def test_cube2hash(self):
     # A good test of i64 math
     self.do_run('// empty file', 'Usage: hashstring <seed>',
-                libraries=self.get_library('third_party/cube2hash', ['libcube2hash.a'], configure=None, make=[EMMAKE, 'make']),
+                libraries=self.get_library('third_party/cube2hash', ['libcube2hash.a'], configure=None),
                 includes=[test_file('third_party/cube2hash')], assert_returncode=NON_ZERO)
 
     for text, output in [('fleefl', '892BDB6FD3F62E863D63DA55851700FDE3ACF30204798CE9'),
@@ -4698,6 +4699,9 @@ Have even and odd!
   def test_printf_octal(self):
     self.do_run_in_out_file_test('printf/test_octal.c')
 
+  def test_printf_macros(self):
+    self.do_core_test('test_printf_macros.c')
+
   def test_vprintf(self):
     self.do_core_test('test_vprintf.c')
 
@@ -5691,9 +5695,10 @@ int main(void) {
   def test_whets(self):
     self.do_runf(test_file('whets.cpp'), 'Single Precision C Whetstone Benchmark')
 
+  # node is slower, and fail on 64-bit
+  @require_v8
   @no_asan('depends on the specifics of memory size, which for asan we are forced to increase')
   def test_dlmalloc_inline(self):
-    self.banned_js_engines = [config.NODE_JS] # slower, and fail on 64-bit
     # needed with typed arrays
     self.set_setting('INITIAL_MEMORY', '128mb')
 
@@ -5701,9 +5706,10 @@ int main(void) {
     self.do_run(src, '*1,0*', args=['200', '1'], force_c=True)
     self.do_run('src.js', '*400,0*', args=['400', '400'], force_c=True, no_build=True)
 
+  # node is slower, and fail on 64-bit
+  @require_v8
   @no_asan('depends on the specifics of memory size, which for asan we are forced to increase')
   def test_dlmalloc(self):
-    self.banned_js_engines = [config.NODE_JS] # slower, and fail on 64-bit
     # needed with typed arrays
     self.set_setting('INITIAL_MEMORY', '128mb')
 
@@ -5956,7 +5962,7 @@ void* operator new(size_t size) {
   def test_lua(self):
     self.emcc_args.remove('-Werror')
 
-    libs = self.get_library('third_party/lua', [Path('src/lua.o'), Path('src/liblua.a')], make=[EMMAKE, 'make', 'generic'], configure=None)
+    libs = self.get_library('third_party/lua', [Path('src/lua.o'), Path('src/liblua.a')], make=['make', 'generic'], configure=None)
     self.do_run('',
                 'hello lua world!\n17\n1\n2\n3\n4\n7',
                 args=['-e', '''print("hello lua world!");print(17);for x = 1,4 do print(x) end;print(10-3)'''],